+++ examples-spring-boot/src/main/java/com/github/dbmdz/flusswerk/spring/boot/example/Application.java
@@ -1,6 +1,6 @@
 package com.github.dbmdz.flusswerk.spring.boot.example;
 
-import com.github.dbmdz.flusswerk.framework.Engine;
+import com.github.dbmdz.flusswerk.framework.engine.Engine;
 import com.github.dbmdz.flusswerk.spring.boot.starter.EnableFlusswerk;
 import com.github.dbmdz.flusswerk.spring.boot.starter.FlusswerkApplication;
 import com.github.dbmdz.flusswerk.spring.boot.starter.FlusswerkProperties;
